backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.8
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.0